requirements.txt
djangorestframework 
djangorestframework-simplejwt 
psycopg2-binary  
python-decouple
requests 
django-filter 